operators/resample: fix issues at the boundaries
Resample will work better if we feed it enough data.

For a minutely series, on a query from `2020-1-1 12:30` to `2020-1-1
14:30`, an hourly resample will likely yield wrong results at the
extremities because it actually needs a whole hour before the left
boundary and after the right boundary. Before this change, we had
values computed on incomplete buckets.

Here we use a rewriter to enlarge the request boundaries by the right
amount, in the scope of the resample operation.

Other similar operators using windowed data might need the same
treatment.