operators/resample: fix the behaviour when upsampling
When upsampling, the provided "freq" argument given to the resample
operator is not what we need: we need the series "freq".

This is done at the cost of an extra computation of the expression
given to resample. A proper query cache is bound to handle this
aspect. It will happen in a not too distant future.

Last note: internally, resample can produce more data than strictly
requested. This is no observed downside in practice but one should
probably be mindful of this.