tsio/interval
Providing reliable intervals for formula is impossible, barring doing
a full .get

We can do a best-guess based on the intervals of the component series.
Such an interval will be "optimistic" or "best case".

Also formulas with the shifted operators will not be handled well.

Support for the autotrophic operators might be turned first-class
(maybe).