test/api: exhibit non-expansion of remote formulas
We don't expand generally because expansion is used by the evaluator
and we want to delegate evaluation to the remote end (for good
reasons: there might be specific operators over there that we don't
know about).

This may annoy some use cases where we want to see a full expansion
for audit or other purposes.