tsio/get: run a loop detection helper to avoid issues with self-referencing formulas
This covers only the simplest case.

More to come, notably mutually referencing formulas (with and without
findseries). However this entails a full graph analysis and cannot be
done at run-time. A dedicated tool will have to be provided. The
probability of this happening also significantly lowers as we add
degrees to the cycle.