types: implement closed vocabularies with Literal
Allow to define enumerated values with Literal
for producing a closed vocabulary (see `tshistory_formula.vocabulary`).
Useful for the formula editor but out of spec formulas
will faill as well when being registered.

The design was in tension on `resample`, `upsample` and `constant`.
The closed vocabulary may be tedious for Pandas power user and
would require a complete migration if all frequencies are built
by  operators.

Note the new requirement of `pytest-golden` for full types validation.

Related to #9